app/requirements.txt
streamlit
requests
langchain
langchain-community
langchain-google-genai
langchain-core
jira
python-dotenv
google-generativeai
testrail-api==1.8.0
pandas
langsmith